follow up r62231, r61779, r62175 * Fix up messages * For new FileRepo::append(), use flags to determine whether to delete or not * Add more error checking for appending * Fix a couple of places in Revision.php and LogPage.php where DB errors were produced when comment was null * Remove bogus checking for !$comment, etc on the DONE phase of chunked uploading * Don't pretend to return a value when raising an exception * Add more tests for chunked uploads * Verify that Status::getErrorsArray() (at least where it is used in ApiUpload::execute()) returns an array that we can pass to dieUsageMessage() * Ensure that checkWarnings(), etc work only on the complete file

+++ includes/filerepo/FSRepo.php
@@ -227,7 +227,7 @@ class FSRepo extends FileRepo {
 		return $status;
 	}
 
-	function append( $srcPath, $toAppendPath ) {
+	function append( $srcPath, $toAppendPath, $flags = 0 ) {
 		$status = $this->newGood();
 
 		// Resolve the virtual URL
@@ -236,20 +236,30 @@ class FSRepo extends FileRepo {
 		}
 		// Make sure the files are there
 		if ( !is_file( $srcPath ) )
-			$status->fatal( 'append-src-filenotfound', $srcPath );
+			$status->fatal( 'filenotfound', $srcPath );
 
 		if ( !is_file( $toAppendPath ) )
-			$status->fatal( 'append-toappend-filenotfound', $toAppendPath );
+			$status->fatal( 'filenotfound', $toAppendPath );
+
+		if ( !$status->isOk() ) return $status;
 
 		// Do the append
-		if( file_put_contents( $srcPath, file_get_contents( $toAppendPath ), FILE_APPEND ) ) {
-			$status->value = $srcPath;
-		} else {
-			$status->fatal( 'fileappenderror', $toAppendPath,  $srcPath);
+		$chunk = file_get_contents( $toAppendPath );
+		if( $chunk === false ) {
+			$status->fatal( 'fileappenderrorread', $toAppendPath );
 		}
 
-		// Remove the source file
-		unlink( $toAppendPath );
+		if( $status->isOk() ) {
+			if ( file_put_contents( $srcPath, $chunk, FILE_APPEND ) ) {
+				$status->value = $srcPath;
+			} else {
+				$status->fatal( 'fileappenderror', $toAppendPath,  $srcPath);
+			}
+		}
+
+		if ( $flags & self::DELETE_SOURCE ) {
+			unlink( $toAppendPath );
+		}
 
 		return $status;
 	}
